refactor(timelock): remove handler helper re-exports
