+++ src/contracts/handlers/default.ts
@@ -9,9 +9,11 @@ import {
 } from "../types";
 import {
     isCsvSpendable,
+} from "./helpers";
+import {
     sequenceToTimelock,
     timelockToSequence,
-} from "./helpers";
+} from "../../utils/timelock";
 
 /**
  * Typed parameters for DefaultVtxo contracts.
