+++ src/contracts/handlers/helpers.ts
@@ -1,23 +1,5 @@
-import { RelativeTimelock } from "../../script/tapscript";
-import {
-    timelockToSequence as timelockToSequenceImpl,
-    sequenceToTimelock as sequenceToTimelockImpl,
-} from "../../utils/timelock";
+import { sequenceToTimelock } from "../../utils/timelock";
 import { Contract, PathContext } from "../types";
-
-/**
- * Convert RelativeTimelock to BIP68 sequence number.
- */
-export function timelockToSequence(timelock: RelativeTimelock): number {
-    return timelockToSequenceImpl(timelock);
-}
-
-/**
- * Convert BIP68 sequence number back to RelativeTimelock.
- */
-export function sequenceToTimelock(sequence: number): RelativeTimelock {
-    return sequenceToTimelockImpl(sequence);
-}
 
 /**
  * Resolve wallet's role from explicit role or by matching pubkey.
