+++ src/contracts/handlers/vhtlc.ts
@@ -11,9 +11,11 @@ import {
     isCltvSatisfied,
     isCsvSpendable,
     resolveRole,
+} from "./helpers";
+import {
     sequenceToTimelock,
     timelockToSequence,
-} from "./helpers";
+} from "../../utils/timelock";
 
 /**
  * Typed parameters for VHTLC contracts.
